move check_amp to checks.py
